requirements.txt
openai
langchain
psutil
pyreadline
prompt_toolkit
colorama
requests
python-dotenv
